backend/requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
python-dotenv==1.0.1
pydantic==2.6.4
google-generativeai==0.5.2
qdrant-client==1.9.0
langchain==0.1.16
langchain-google-genai==1.0.3
httpx==0.27.0
python-multipart==0.0.9
